requirements.txt
google-adk
google-generativeai
google-genai
fastapi
uvicorn
google-cloud-firestore
google-cloud-datastore
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
nest_asyncio